User commit

+++ src/main/java/com/example/demo/entity/StudentEntity.java
@@ -1,32 +1,41 @@
 package com.example.demo.entity;
 
-public class StudentEntity{
-    private int id;
+public class StudentEntity {
+
+    private Long id;  
     private String name;
     private String email;
-    public StudentEntity(){
+
+    public StudentEntity() {
     }
-    public StudentEntity(int id,String name,String email){
-        this.id=id;
-        this.name=name;
-        this.email=email;
+
+    public StudentEntity(Long id, String name, String email) { 
+        this.id = id;
+        this.name = name;
+        this.email = email;
     }
-    public int getId(){
+
+    public Long getId() {      
         return id;
     }
-    public void setId(int id){
-        this.id=id;
+
+    public void setId(Long id) {   
+        this.id = id;
     }
-    public String getName(){
+
+    public String getName() {
         return name;
     }
-    public void setName(String name){
-        this.name=name;
+
+    public void setName(String name) {
+        this.name = name;
     }
-    public String getEmail(){
+
+    public String getEmail() {
         return email;
     }
-    public void setEmail(String email){
-        this.email=email;
+
+    public void setEmail(String email) {
+        this.email = email;
     }
 }
